THE SHIT
So basically, we now have a player who spawns in a random area in the grid with 5 random bombs, and if u use wasd or arrow keys, the play will move. If they try to go beyond the border, you will hear my fabulous voice. DONT SPAM IT UR PC WILL CRASH! What we need now is to implement the sound to continuously play depending on the range from a bomb, and also we need to calculate how far away a bomb is. After that, we have a game!!!